Implement the backbone of the website in HTML by adding a home page and some sample recipe pages
Created three files:
    - index.html
    - sugar_cookies.html
    - brownies.html

The first of these is the home page, the last two are pages containing recipes for sugar cookies and
brownies, respectively. Images were added for both recipes on their respective pages, alongside a list
for the ingredients and steps on how to proceed with the recipe. A brief description proceeds the cooking
instructions.